Move refresh_access_token to new apps.users.utils module Implement refresh_access_token during logout
